tied in radio buttons with text input
