+++ src/components/Quizzes/CreateQuiz.js
@@ -5,133 +5,27 @@ import {Question} from './Question'
 import CategoriesRepository from  '../../data/CategoriesRepository'
 import { QuestionContext } from './QuestionContext'
 import UserRepository from '../../data/UserRepository'
-
+import { QuizForm } from './QuizForm'
 
 
 
 export const CreateQuiz = () => {
 
-    const [quizData, updateQuiz] = useState({
-        quizId: 0,
-        quizName: "",
-        categoryId: "",
-        userId: 0
-    })
-
-    const [quizName, setQuizName] = useState("")
-    const [currentQuestion, setCurrentQuestion] = useState(0)
-    const [questionArray, updateQuestionArray] = useState([ { number: 1 }])
-    const [categories, updateCategories] = useState([])
-    const [quizCategory, setQuizCategory] = useState("")
-    const [currentUser, setCurrentUser] = useState("")
-    const history = useHistory()
+    const [quizForm, toggleQuizForm] = useState(true)
 
 
-    useEffect(() => {
-        CategoriesRepository.getAllCategories()
-        .then(r => updateCategories(r))
-        // .then(UserRepository.getCurrentUser())
-        // .then(r => setCurrentUser(r))
-        // .then(console.log(currentUser))
-    }, []
-    )
+// POST UP quiz, get the category and quiz name and user Id
+// THEN fetch it back so you can assign it to those questions.
+// const handleCreateQuiz {
 
-
-
-    const addQuestion = () => {
-        // Before adding a question, post Quiz to database?
-        // If Quiz is canceled, DELETE from database.
-        // When submitting, assign all those questions for the quiz to the QuizId in database.
-
-
-        const newQuestion = {
-            number: questionArray.length + 1,
-            quizId: 0
-            // questionTypeId:
-            // content:
-        }
-
-
-        // Fill out all data before addquestion can be pressed
-        // Post question when 'Addquestion is pressed'
-        // Fetch the question array
-        updateQuestionArray(questionArray => [...questionArray, newQuestion])
-
-    }
-
-    const handleCancel = () => {
-        history.push("/")
-    }
-
-    const handleQuizSubmit = (e) => {
-
-        e.preventDefault()
-
-        const newQuiz = {
-            quizName: `${quizData.quizName}`,
-            categoryId: quizData.categoryId,
-            userId: quizData.userId
-        }
-
-        updateQuiz(newQuiz)
-
-    }
-
+// }
 
     return (
         <>
 
-            <form className="quizForm">
-            <h1>Create a Quiz</h1>
-                <div className="quiz-form"></div>
-                <label htmlFor="quizName">Quiz Name:    </label>
-                <input
-                    type="text"
-                    required
-                    autoFocus
-                    className="form-control"
-                    onChange={e => setQuizName(e.target.value)}
-                    id="quizName"
-                    placeholder="Quiz name"
-                />
-                <div className="form--categories">
-                <label htmlFor="categories">Category:</label>
-                <br></br>
-                {categories.map((category) => { return (
-                    <>
-                    <label htmlFor={`category--${category.name}`}>{`${category.name}`}</label>
-                    <input
-                        type="radio"
-                        required
-                        className="form-control"
-                        id={`category--${category.name}`}
-                        name="categoryGroup"
-                        value={category.id}
-                        onChange={e => setQuizCategory(e.target.value)}
-                    />
-                    <br></br>
-                    </>)
-                })}
-                </div>
-                {/* Generate multiple questions */}
-                <QuestionContext.Provider value = {[questionArray, updateQuestionArray]}>
-                    {/* Map over question array */}
-                {questionArray.map(
-                    question => {
-                        return <Question number={question.number}/>
-                    }
-                )}
-                </QuestionContext.Provider>
-                <div className="addQuestion">
-                    <button onClick={() => addQuestion()}>Add Question</button>
-                </div>
-                <div className="submitButton">
-                    <button>Submit Quiz</button>
-                </div>
-                <div className="cancelButton">
-                    <button onClick={() => handleCancel()}>Cancel</button>
-                </div>
-            </form>
+            <button>Create New Quiz</button>
+            <QuizForm toggled = {quizForm} />
+
         </>
     )
 }
